- 05.12.17 CHG : ssiml/ISUnderSample : ignores empty classes : Johannes - 06.12.17 CHG : ssiml/Annotation : skips labels with NaN score when extracting a sample list : Johannes - 06.12.17 CHG : ssiml/Annotation : support for completing continuous annotations : Johannes - 07.12.17 CHG : ssicml/CMLAnnotation : predicts scores within scheme range : Johannes - 19.12.17 CHG : ssiml/Annotation : added convert_to_stream function : Johannes - 19.12.17 CHG : cmltrain : added --convert_anno_to_stream switch : Johannes - 19.12.17 CHG : cmltrain : renamed --convert to --convert_stream_to_anno switch : Johannes - 08.01.17 CHG : IModel : added getModelType() and confidence value in forward method : Johannes - 08.01.17 CHG : cmltrain : forward confidence : Johannes

+++ core/include/ioput/socket/osc/OscTypes.h
@@ -1,35 +1,36 @@
 /*
-	oscpack -- Open Sound Control packet manipulation library
-	http://www.audiomulch.com/~rossb/oscpack
+    oscpack -- Open Sound Control packet manipulation library
+    http://www.audiomulch.com/~rossb/oscpack
 
-	Copyright (c) 2004-2005 Ross Bencina <rossb@audiomulch.com>
+    Copyright (c) 2004-2005 Ross Bencina <rossb@audiomulch.com>
 
-	Permission is hereby granted, free of charge, to any person obtaining
-	a copy of this software and associated documentation files
-	(the "Software"), to deal in the Software without restriction,
-	including without limitation the rights to use, copy, modify, merge,
-	publish, distribute, sublicense, and/or sell copies of the Software,
-	and to permit persons to whom the Software is furnished to do so,
-	subject to the following conditions:
+    Permission is hereby granted, free of charge, to any person obtaining
+    a copy of this software and associated documentation files
+    (the "Software"), to deal in the Software without restriction,
+    including without limitation the rights to use, copy, modify, merge,
+    publish, distribute, sublicense, and/or sell copies of the Software,
+    and to permit persons to whom the Software is furnished to do so,
+    subject to the following conditions:
 
-	The above copyright notice and this permission notice shall be
-	included in all copies or substantial portions of the Software.
+    The above copyright notice and this permission notice shall be
+    included in all copies or substantial portions of the Software.
 
-	Any person wishing to distribute modifications to the Software is
-	requested to send the modifications to the original developer so that
-	they can be incorporated into the canonical version.
+    Any person wishing to distribute modifications to the Software is
+    requested to send the modifications to the original developer so that
+    they can be incorporated into the canonical version.
 
-	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-	EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-	MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-	IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR
-	ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
-	CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-	WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+    EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+    MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+    IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR
+    ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
+    CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+    WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 #ifndef INCLUDED_OSCTYPES_H
 #define INCLUDED_OSCTYPES_H
 
+#include <stdint.h>
 
 namespace ssi {
 
@@ -42,8 +43,8 @@ typedef unsigned __int64 osc_uint64;
 
 #else
 
-typedef long long osc_int64;
-typedef unsigned long long osc_uint64;
+typedef int64_t osc_int64;
+typedef uint64_t osc_uint64;
 
 #endif
 
@@ -51,13 +52,13 @@ typedef unsigned long long osc_uint64;
 
 #ifdef x86_64
 
-typedef signed int osc_int32;
-typedef unsigned int osc_uint32;
+typedef int32_t osc_int32;
+typedef uint32_t osc_uint32;
 
 #else
 
-typedef signed long osc_int32;
-typedef unsigned long osc_uint32;
+typedef int32_t osc_int32;
+typedef uint32_t osc_uint32;
 
 #endif
